topic suppport replica_num

+++ tencentcloud/resource_tc_ckafka_topic.go
@@ -253,8 +253,7 @@ func resourceTencentCloudCkafkaTopicRead(d *schema.ResourceData, meta interface{
 	}
 	items := strings.Split(d.Id(), FILED_SP)
 	if len(items) < 2 {
-		d.SetId("")
-		return nil
+		return fmt.Errorf("id is broken,%s", d.Id())
 	}
 	instanceId := items[0]
 	topicName := items[1]
@@ -314,9 +313,16 @@ func resourceTencentCloudCkafkaTopicUpdate(d *schema.ResourceData, meta interfac
 	ckafkcService := CkafkaService{
 		client: meta.(*TencentCloudClient).apiV3Conn,
 	}
+
+	items := strings.Split(d.Id(), FILED_SP)
+	if len(items) < 2 {
+		return fmt.Errorf("id is broken,%s", d.Id())
+	}
+	instanceId := items[0]
+	topicName := items[1]
+
 	request := ckafka.NewModifyTopicAttributesRequest()
-	instanceId := d.Get("instance_id").(string)
-	topicName := d.Get("topic_name").(string)
+	replicaNum := d.Get("replica_num").(int)
 	whiteListSwitch := d.Get("enable_white_list").(bool)
 	cleanUpPolicy := d.Get("clean_up_policy").(string)
 	retention := d.Get("retention").(int)
@@ -332,6 +338,7 @@ func resourceTencentCloudCkafkaTopicUpdate(d *schema.ResourceData, meta interfac
 	}
 	request.InstanceId = &instanceId
 	request.TopicName = &topicName
+	request.ReplicaNum = helper.IntInt64(replicaNum)
 	request.EnableWhiteList = helper.BoolToInt64Ptr(whiteListSwitch)
 	request.MinInsyncReplicas = helper.IntInt64(d.Get("sync_replica_min_num").(int))
 	request.UncleanLeaderElectionEnable = helper.BoolToInt64Ptr(d.Get("unclean_leader_election_enable").(bool))
@@ -393,8 +400,12 @@ func resourceTencentCLoudCkafkaTopicDelete(d *schema.ResourceData, meta interfac
 	ckafkcService := CkafkaService{
 		client: meta.(*TencentCloudClient).apiV3Conn,
 	}
-	instanceId := d.Get("instance_id").(string)
-	topicName := d.Get("topic_name").(string)
+	items := strings.Split(d.Id(), FILED_SP)
+	if len(items) < 2 {
+		return fmt.Errorf("id is broken,%s", d.Id())
+	}
+	instanceId := items[0]
+	topicName := items[1]
 
 	err := ckafkcService.DeleteCkafkaTopic(ctx, instanceId, topicName)
 	if err != nil {
